Optimize product deletion process
The product deletion method in the admin area has been improved. An HTTP Delete attribute was added to enhance clarity and enforce correct usage. The image deletion process was also refined: now, it first checks if the image URL is not empty before attempting to delete the file. This prevents unnecessary path construction and file existence checks for products without images, improving efficiency.

+++ WarlockMVC/Areas/Admin/Controllers/ProductController.cs
@@ -125,6 +125,7 @@ namespace WarlockMVC.Areas.Admin.Controllers
             return Json(new { data = objProductList });
         }
 
+        [HttpDelete]
         public IActionResult Delete(int? id)
         {
             var productToBeDeleted = _unitOfWork.Product.Get(x => x.Id == id);
@@ -133,14 +134,17 @@ namespace WarlockMVC.Areas.Admin.Controllers
                 return Json(new { success = false, message = "Error while deleting" });
             }
 
-            var oldImagePath = Path.Combine(
-                _webHostEnvironment.WebRootPath,
-                productToBeDeleted.ImageUrl.TrimStart('\\')
-            );
-
-            if (System.IO.File.Exists(oldImagePath))
+            if (!string.IsNullOrEmpty(productToBeDeleted.ImageUrl))
             {
-                System.IO.File.Delete(oldImagePath);
+                var oldImagePath = Path.Combine(
+                    _webHostEnvironment.WebRootPath,
+                    productToBeDeleted.ImageUrl.TrimStart('\\')
+                );
+
+                if (System.IO.File.Exists(oldImagePath))
+                {
+                    System.IO.File.Delete(oldImagePath);
+                }
             }
 
             _unitOfWork.Product.Delete(productToBeDeleted);
